+++ src/cellTypes/latex.ts
@@ -8,6 +8,7 @@ import { cellControlsTemplate } from "../components/controls";
 import { StarboardTextEditor } from "../components/textEditor";
 import { Cell, CellElements, CellHandlerAttachParameters, ControlButton, Runtime } from "../types";
 import { katexLoader } from "../components/helpers/katex";
+import { dispatchStarboardEvent } from "../components/helpers/event";
 
 export const LATEX_CELL_TYPE_DEFINITION = {
   name: "LateX (KaTeX)",
@@ -31,7 +32,7 @@ export class LatexCellHandler extends BaseCellHandler {
       editOrRunButton = {
         icon: "bi bi-play-circle",
         tooltip: "Render LaTeX",
-        callback: () => this.runtime.controls.emit({ id: this.cell.id, type: "RUN_CELL" }),
+        callback: (evt, dispatch) => dispatch("sb:run_cell", { id: this.cell.id }),
       };
     } else {
       editOrRunButton = {